add Spark DB engine spec
