+++ superset/db_engine_specs/hive.py
@@ -580,14 +580,4 @@ class HiveEngineSpec(PrestoEngineSpec):
         """
 
         return True
-
-
-class SparkEngineSpec(HiveEngineSpec):
-
-    engine_name = "Apache Spark SQL"
-
-
-class Spark3EngineSpec(HiveEngineSpec):
-
-    engine_name = "Apache Spark3 SQL"
